[opentitantool] move process's stdout printer to its own module
It was previously used to log the output of verilator using the
logging system. This commit moves the printer to its own module
and allows the caller to specify the logging target. With the
current logger's configuration, this target is not used (yet).
