+++ sw/host/opentitanlib/src/transport/verilator/mod.rs
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod gpio;
-mod stdout;
 pub mod subprocess;
 pub mod transport;
 pub mod uart;
